Add files via upload
Forecasting is a method of predicting the future using the data from the past, this helps different forums to take informed decisions by determining the future trends by predicting using the estimates. By comparing different forecasting methods RMSE and observing the graph, it can be deduced that the drift method has a comparatively smaller confidence interval and RMSE, making it the best method.
Now, to predict the values for a future time interval we use ARIMA that stands for AutoReegressive Integrated Moving Average, this forecasts the future values based on past events. 
Finally, as PCE is a major component of the GDP, continuous and systematic forecasting will help in enabling efficient policies for the citizens. This data is useful for econometric scientists and policy makers to make cutting-edge decisions for the government.